Move `DEBUG` code paths from the main `SceneDelegate` to a new `DebuggingSceneDelegate` class to separate test specific code from production
